+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `TileScopePack.make_tracked` return a new pack with tracked fusion instead of
   only adding the `AddAssumptionFactory` to the pack
 - `make_fusion` and `make_interleaving` make the pack name a little more descriptive
+- Updated short obstruction verification to take a length argument
 
 ## [2.4.1] - 2020-10-28
 ### Added